. JavaRpcService.java, proxy.i, rpcbase.i, server.i, Helpers.java, SimpFileSvcsvr.java, StreamSvcsvr.java, maincli.java: fixed bugs to make the java version of fulltest work.
. genjava.cpp: added setting property propNoReply to no-reply's request
parameters to send, which was missed and caused the response returns.

. iftasks.cpp: add check to CIfInvokeMethodTask's OnTaskComplete whether the
request has response by HasReply, as depended only on the existance of
propRespPtr, which was not adaquate when the CIfInvokeMethodTask is canceled,
and the response was set by the CIfInvokeMethodTask's OnCancel.

. dbusprxy.cpp: added CF_WITH_REPLY flag to the call options as was mandatory
due to the change in CIfInvokeMethodTask.

. rpcif.cpp: fixed a bug in DoInvoke_SendData, which set the call option
property to the interface object, rather than the callback object( the
invokeMethod task ), which was not significant till the change in
CIfInvokeMethodTask.

. rpcif.cpp: commented out the CIfSvrConnMgr, since sometimes it falls in a
weird state, that the response cannot be sent because the IsConnected( szAddr )
reported false while the interface object is well connected.

+++ ridl/javatest/org/rpcf/fulltest/SimpFileSvcsvr.java
@@ -95,7 +95,7 @@ public class SimpFileSvcsvr extends SimpFileSvcsvrbase
         String szText )
     {
         oReqCtx.setResponse( RC.STATUS_SUCCESS,
-                i1, i2, i3, i4 + .1, i5 +.2, szText );
+                i1, i2, i3, (float)(i4 + .1), i5 +.2, szText );
         return RC.STATUS_SUCCESS;
     }
     public void onEchoManyCanceled(
